requirements.txt
deep-translator==1.11.4
translators==5.9.1
huggingface-hub==0.22.2
manga-ocr==0.1.11
numpy==1.24.2
opencv-python==4.9.0.80
pillow==10.3.0
ultralytics==8.1.43
Flask==2.2.5
flask-socketio==5.3.6
python-socketio==5.10.0
python-engineio==4.8.1
Werkzeug==3.0.1
gunicorn==21.2.0
tqdm==4.66.2
safetensors==0.4.2
cryptography==42.0.4
torch==2.0.1
torchvision==0.15.2
sentencepiece==0.2.0
chrome-lens-py>=3.0.0
google-generativeai>=0.3.0